# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  vigia-verde-go_internal_application_event.CreateDTO:
    properties:
      description:
        type: string
      event_type:
        type: string
      image_src:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      title:
        type: string
    required:
    - event_type
    - latitude
    - longitude
    - title
    type: object
  vigia-verde-go_internal_domain_event.Author:
    properties:
      id:
        type: string
      name:
        type: string
      subscribed:
        type: boolean
    type: object
  vigia-verde-go_internal_domain_event.Event:
    properties:
      author:
        $ref: '#/definitions/vigia-verde-go_internal_domain_event.Author'
      comment_count:
        type: integer
      comments:
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      downvotes:
        type: integer
      event_type:
        $ref: '#/definitions/vigia-verde-go_internal_domain_event.EventType'
      geohash:
        type: string
      id:
        type: string
      image_src:
        type: string
      location:
        $ref: '#/definitions/vigia-verde-go_internal_domain_event.GeoPoint'
      title:
        type: string
      upvotes:
        type: integer
    type: object
  vigia-verde-go_internal_domain_event.EventType:
    enum:
    - removal
    - pruning
    - seedling
    type: string
    x-enum-varnames:
    - EventRemoval
    - EventPruning
    - EventSeedling
  vigia-verde-go_internal_domain_event.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
host: https://vigia-verde-514678222430.southamerica-east1.run.app
info:
  contact:
    name: Arthur de Oliveira & Italo de Oliveira
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Vigia Verde API Docs
  version: 0.1.0
paths:
  /events:
    get:
      description: Retorna uma lista de eventos com suporte a paginação, filtros por
        autor/tipo e busca geoespacial por coordenadas.
      parameters:
      - description: Filtrar por ID do autor
        in: query
        name: author_id
        type: string
      - description: Filtrar por tipo de evento
        in: query
        name: event_type
        type: string
      - description: Número da página (padrão capturado por web.GetPagination)
        in: query
        name: page
        type: integer
      - description: Limite de itens por página
        in: query
        name: limit
        type: integer
      - description: Latitude para busca geoespacial
        in: query
        name: lat
        type: number
      - description: Longitude para busca geoespacial
        in: query
        name: lng
        type: number
      - description: Precisão da busca (capturado por web.GetPrecision)
        in: query
        name: precision
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Nota: O web.Respond envia a lista e o meta de paginação'
          schema:
            items:
              $ref: '#/definitions/vigia-verde-go_internal_domain_event.Event'
            type: array
        "400":
          description: Erro nos parâmetros enviados
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar eventos
      tags:
      - Eventos
    post:
      consumes:
      - application/json
      description: Registra um novo evento a partir dos dados fornecidos no corpo
        da requisição.
      parameters:
      - description: Dados para criação do evento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vigia-verde-go_internal_application_event.CreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 'Exemplo: {"id": "123"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid json
          schema:
            type: string
        "500":
          description: Erro interno tratado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar um evento
      tags:
      - Eventos
  /events/{id}:
    get:
      description: Retorna os detalhes completos de um único evento baseado no parâmetro
        de URL {id}.
      parameters:
      - description: ID do Evento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vigia-verde-go_internal_domain_event.Event'
        "404":
          description: Evento não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Buscar um evento por ID
      tags:
      - Eventos
swagger: "2.0"